Issue-211: Moved BaseFragment to core module.
